requirements.txt
pandas
faker
python-dotenv
supabase
psycopg2-binary